refactor: move securing into LocalSecureManager, dedup names, guard restore
move moveMediaToSecureFolder out of ImageFunctions into LocalSecureManager
alongside the rest of the secure logic.

give each secured file a unique destination so two sources with the same
basename can't overwrite each other's ciphertext; reuse an original's
existing secured path on re-secure so we never orphan a row-less file.

recover a corrupted file iv in toSecureMedia, and refuse to restore an
encrypted item whose iv can't be recovered instead of raw-copying the
ciphertext into the gallery and deleting the only copy.
